merubah file class obat menjadi file class kategori . mejadi kan class obat ke class kategori agar bisa di jadik kan class parent. agar bisa di pakai di class child.
